refactor code and unit test
useless import

refactor
