+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/HiveTableScanSuite.scala
@@ -17,8 +17,6 @@
 
 package org.apache.spark.sql.hive.execution
 
-import org.apache.spark.SparkContext
-import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.hive.test.TestHive
 import org.apache.spark.sql.hive.test.TestHive._
@@ -91,15 +89,11 @@ class HiveTableScanSuite extends HiveComparisonTest {
     assert(sql("select casesensitivecolname from spark_4959_2").head() === Row("hi"))
   }
 
-  test("combine small files") {
-    def getPartitions[T](sc: SparkContext, rdd: RDD[T]): Array[Long] = {
-      sc.runJob(rdd, (x: Iterator[T]) => 1)
-    }
+  test("Spark-8813 Combine small splits for table scan") {
     val partitionNum = 5
     val partitionTable = "combine_small"
     sql("set hive.exec.dynamic.partition.mode=nonstrict")
-    sql("set spark.sql.mapper.splitCombineSize=1000000")
-    val df = (1 to 100).map { i => (i, i) }.toDF("a", "b").repartition(100)
+    val df = (1 to 100).map { i => (i, i) }.toDF("a", "b").coalesce(100)
     df.registerTempTable("temp")
     sql(
       s"""create table $partitionTable (a int, b string)
@@ -108,13 +102,18 @@ class HiveTableScanSuite extends HiveComparisonTest {
     sql(
       s"""insert into table $partitionTable partition(c)
           |select a, b, (b % $partitionNum) as c from temp""".stripMargin)
-    val result = sql( s"""select * from $partitionTable order by a""")
-    val rddPartitions = getPartitions(TestHive.sparkContext, result.rdd)
-    // Ensure that there are only have 4 RDD partitions after combination
-    assert(rddPartitions.length == 5)
-    // Ensure that the result is the same as the original after combination
+
+    // Check the num of RDD partition without the combination
+    assert(sql( s"""select * from $partitionTable""").rdd.getNumPartitions == 100)
+
+    // Check the num of RDD partitions with the combination
+    sql("set spark.sql.mapper.splitCombineSize=10000")
+    assert(sql(
+      s"""select * from $partitionTable""").rdd.getNumPartitions == partitionNum)
+
+    // Ensure that the result is the same as the original after the combination
     assert(
-      result.collect().map(_.toString()).deep
+      sql( s"""select * from $partitionTable order by a""").collect().map(_.toString()).deep
       ==  (1 to 100).map{i => s"[$i,$i,${i % partitionNum}]"}.toArray.deep
     )
   }
